# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for personal blog.",
        "title": "Golang Restful API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/article": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Get list of articles.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/example.GetArticles"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "\u003c/p\u003e[author_id] : Author id (min:1)\u003c/p\u003e[title] : The article title \u003c/p\u003e[desc] : Description of article \u003c/p\u003e[content] : Content of article \u003c/p\u003e[imageUrl] : Url of image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Add new article",
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/example.PostArticle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    }
                }
            }
        },
        "/article/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Get article info by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/example.GetArticle"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.ArticleNotExist"
                        }
                    }
                }
            },
            "put": {
                "description": "\u003c/p\u003e[author_id] : Author id (min:1)\u003c/p\u003e[title] : The article title \u003c/p\u003e[desc] : Description of article \u003c/p\u003e[content] : Content of article \u003c/p\u003e[imageUrl] : Url of image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Modify article info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/example.PostArticle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.ArticleNotExist"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Delete article info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.ArticleNotExist"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "example.ArticleInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "object",
                    "$ref": "#/definitions/example.AuthorInfo"
                },
                "author_id": {
                    "type": "integer",
                    "example": 1
                },
                "content": {
                    "type": "string",
                    "example": "Although classified as a comedy...."
                },
                "created_on": {
                    "type": "integer",
                    "example": 1574952062
                },
                "deleted_on": {
                    "type": "integer",
                    "example": 0
                },
                "desc": {
                    "type": "string",
                    "example": "The Merchant of Venice is a 16th-century play written by William Shakespeare"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "http://xxx.xxx.xxx.jpeg"
                },
                "modified_on": {
                    "type": "integer",
                    "example": 1575189888
                },
                "title": {
                    "type": "string",
                    "example": "The Merchant Venice "
                }
            }
        },
        "example.ArticleNotExist": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 10003
                },
                "error_message": {
                    "type": "string",
                    "example": "Article not exist"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "example.AuthorInfo": {
            "type": "object",
            "properties": {
                "born": {
                    "type": "integer",
                    "example": 1564
                },
                "created_on": {
                    "type": "integer",
                    "example": 1574952062
                },
                "deleted_on": {
                    "type": "integer",
                    "example": 0
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "modified_on": {
                    "type": "integer",
                    "example": 1575189888
                },
                "name": {
                    "type": "string",
                    "example": "William Shakespeare"
                }
            }
        },
...
        "example.GetArticle": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 200
                },
                "error_message": {
                    "type": "string",
                    "example": "OK"
                },
                "result": {
                    "type": "object",
                    "$ref": "#/definitions/example.ArticleInfo"
                }
            }
        },
        "example.GetArticles": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 200
                },
                "error_message": {
                    "type": "string",
                    "example": "OK"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/example.ArticleInfo"
                    }
                }
            }
        },
...
        "example.OK": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 200
                },
                "error_message": {
                    "type": "string",
                    "example": "OK"
                },
                "result": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "example.PostArticle": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer",
                    "example": 1
                },
                "content": {
                    "type": "string",
                    "example": "Although classified as a comedy...."
                },
                "desc": {
                    "type": "string",
                    "example": "The Merchant of Venice is a 16th-century play written by William Shakespeare"
                },
                "image_url": {
                    "type": "string",
                    "example": "http://xxx.xxx.xxx.jpeg"
                },
                "title": {
                    "type": "string",
                    "example": "The Merchant Venice "
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}